Safari-609.1.11.1/LayoutTests/platform/win/fast/dom/html-link-element-activation-behavior-url-is-null-expected.txt

CONSOLE MESSAGE: line 13: TypeError: MouseEventConstructor is not a constructor (evaluating 'new MouseEvent("click")')
This tests html link element activation behavior when url is null

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL successfullyParsed should be true (of type boolean). Was undefined (of type undefined).

TEST COMPLETE

